circle/requirements.txt
PyYAML==3.12
ruamel.yaml==0.15.40
